# Fix performance issues in R 4.0.0 with an `owned` argument for assignment