--- a/R/slice.R
+++ b/R/slice.R
@@ -183,8 +183,8 @@ vec_assign_seq <- function(x, value, start, size, increasing = TRUE) {
   .Call(vctrs_assign_seq, x, value, start, size, increasing)
 }
 
-vec_assign_params <- function(x, i, value, assign_names = FALSE) {
-  .Call(vctrs_assign_params, x, i, value, assign_names)
+vec_assign_params <- function(x, i, value, assign_names = FALSE, owned = FALSE) {
+  .Call(vctrs_assign_params, x, i, value, assign_names, owned)
 }
 
 vec_remove <- function(x, i) {