--- a/src/slice-assign-array.c
+++ b/src/slice-assign-array.c
@@ -1,9 +1,16 @@
 #include "vctrs.h"
 #include "utils.h"
 #include "strides.h"
+#include "slice-assign.h"
 
 #define ASSIGN_SHAPED_INDEX(CTYPE, DEREF, CONST_DEREF)   \
-  SEXP out = PROTECT(r_maybe_duplicate(proxy));          \
+  SEXP out;                                              \
+  if (owned) {                                           \
+    out = PROTECT(r_maybe_duplicate_shared(proxy));      \
+  } else {                                               \
+    out = PROTECT(r_maybe_duplicate(proxy));             \
+  }                                                      \
+                                                         \
   CTYPE* p_out = DEREF(out);                             \
                                                          \
   const CTYPE* p_value = CONST_DEREF(value);             \
@@ -35,7 +42,13 @@
   return out
 
 #define ASSIGN_SHAPED_COMPACT(CTYPE, DEREF, CONST_DEREF) \
-  SEXP out = PROTECT(r_maybe_duplicate(proxy));          \
+  SEXP out;                                              \
+  if (owned) {                                           \
+    out = PROTECT(r_maybe_duplicate_shared(proxy));      \
+  } else {                                               \
+    out = PROTECT(r_maybe_duplicate(proxy));             \
+  }                                                      \
+                                                         \
   CTYPE* p_out = DEREF(out);                             \
                                                          \
   const R_len_t start = p_info->p_index[0];              \
@@ -73,22 +86,22 @@
     ASSIGN_SHAPED_INDEX(CTYPE, DEREF, CONST_DEREF);   \
   }
 
-static inline SEXP lgl_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP lgl_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_SHAPED(int, LOGICAL, LOGICAL_RO);
 }
-static inline SEXP int_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP int_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_SHAPED(int, INTEGER, INTEGER_RO);
 }
-static inline SEXP dbl_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP dbl_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_SHAPED(double, REAL, REAL_RO);
 }
-static inline SEXP cpl_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP cpl_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_SHAPED(Rcomplex, COMPLEX, COMPLEX_RO);
 }
-static inline SEXP chr_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP chr_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_SHAPED(SEXP, STRING_PTR, STRING_PTR_RO);
 }
-static inline SEXP raw_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP raw_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_SHAPED(Rbyte, RAW, RAW_RO);
 }
 
@@ -99,7 +112,13 @@ static inline SEXP raw_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct 
 // -----------------------------------------------------------------------------
 
 #define ASSIGN_BARRIER_SHAPED_INDEX(GET, SET)            \
-  SEXP out = PROTECT(r_maybe_duplicate(proxy));          \
+  SEXP out;                                              \
+  if (owned) {                                           \
+    out = PROTECT(r_maybe_duplicate_shared(proxy));      \
+  } else {                                               \
+    out = PROTECT(r_maybe_duplicate(proxy));             \
+  }                                                      \
+                                                         \
   R_len_t k = 0;                                         \
                                                          \
   for (R_len_t i = 0; i < p_info->shape_elem_n; ++i) {   \
@@ -129,7 +148,12 @@ static inline SEXP raw_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct 
   return out
 
 #define ASSIGN_BARRIER_SHAPED_COMPACT(GET, SET)         \
-  SEXP out = PROTECT(r_maybe_duplicate(proxy));         \
+  SEXP out;                                             \
+  if (owned) {                                          \
+    out = PROTECT(r_maybe_duplicate_shared(proxy));     \
+  } else {                                              \
+    out = PROTECT(r_maybe_duplicate(proxy));            \
+  }                                                     \
                                                         \
   const R_len_t start = p_info->p_index[0];             \
   const R_len_t n = p_info->p_index[1];                 \
@@ -166,7 +190,7 @@ static inline SEXP raw_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct 
     ASSIGN_BARRIER_SHAPED_INDEX(GET, SET);   \
   }
 
-static SEXP list_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static SEXP list_assign_shaped(SEXP proxy, SEXP index, SEXP value, bool owned, struct strides_info* p_info) {
   ASSIGN_BARRIER_SHAPED(VECTOR_ELT, SET_VECTOR_ELT);
 }
 
@@ -176,15 +200,19 @@ static SEXP list_assign_shaped(SEXP proxy, SEXP index, SEXP value, struct stride
 
 // -----------------------------------------------------------------------------
 
-static inline SEXP vec_assign_shaped_switch(SEXP proxy, SEXP index, SEXP value, struct strides_info* p_info) {
+static inline SEXP vec_assign_shaped_switch(SEXP proxy,
+                                            SEXP index,
+                                            SEXP value,
+                                            bool owned,
+                                            struct strides_info* p_info) {
   switch (vec_proxy_typeof(proxy)) {
-  case vctrs_type_logical:   return lgl_assign_shaped(proxy, index, value, p_info);
-  case vctrs_type_integer:   return int_assign_shaped(proxy, index, value, p_info);
-  case vctrs_type_double:    return dbl_assign_shaped(proxy, index, value, p_info);
-  case vctrs_type_complex:   return cpl_assign_shaped(proxy, index, value, p_info);
-  case vctrs_type_character: return chr_assign_shaped(proxy, index, value, p_info);
-  case vctrs_type_raw:       return raw_assign_shaped(proxy, index, value, p_info);
-  case vctrs_type_list:      return list_assign_shaped(proxy, index, value, p_info);
+  case vctrs_type_logical:   return lgl_assign_shaped(proxy, index, value, owned, p_info);
+  case vctrs_type_integer:   return int_assign_shaped(proxy, index, value, owned, p_info);
+  case vctrs_type_double:    return dbl_assign_shaped(proxy, index, value, owned, p_info);
+  case vctrs_type_complex:   return cpl_assign_shaped(proxy, index, value, owned, p_info);
+  case vctrs_type_character: return chr_assign_shaped(proxy, index, value, owned, p_info);
+  case vctrs_type_raw:       return raw_assign_shaped(proxy, index, value, owned, p_info);
+  case vctrs_type_list:      return list_assign_shaped(proxy, index, value, owned, p_info);
   default: Rf_error("Internal error: Non-vector type `%s` in `vec_assign_shaped_switch()`",
                     vec_type_as_str(vec_proxy_typeof(proxy)));
   }
@@ -193,13 +221,13 @@ static inline SEXP vec_assign_shaped_switch(SEXP proxy, SEXP index, SEXP value, 
 // -----------------------------------------------------------------------------
 
 // [[ include("vctrs.h") ]]
-SEXP vec_assign_shaped(SEXP proxy, SEXP index, SEXP value) {
+SEXP vec_assign_shaped(SEXP proxy, SEXP index, SEXP value, const struct vec_assign_opts* opts) {
   int n_protect = 0;
 
   struct strides_info info = new_strides_info(proxy, index);
   PROTECT_STRIDES_INFO(&info, &n_protect);
 
-  SEXP out = vec_assign_shaped_switch(proxy, index, value, &info);
+  SEXP out = vec_assign_shaped_switch(proxy, index, value, opts->owned, &info);
 
   UNPROTECT(n_protect);
   return out;