… src/utils.h
… src/vctrs.h
… tests/testthat/helper-memory.R
… tests/testthat/test-bind.R
… tests/testthat/test-slice-assign.R