Closes #1057 

After speaking with @lionel- about this, it seems like this is an alternate approach to fixing the performance issues seen in R 4.0.0.

@lionel- also reminded me that this is very similar to the old way `vec_assign()` used to work, with a `clone` parameter.

This approach uses an `owned` argument to decide how to _potentially_ duplicate the proxy. The core change is:

```c
if (owned) {
  out = r_maybe_duplicate_shared(x);
} else {
  out = r_maybe_duplicate(x);
}
```

Where:
- `r_maybe_duplicate_shared()` duplicates if `refcount(x) > 1`
- `r_maybe_duplicate()` duplicates if `refcount(x) > 0`

This distinction matters on R 4.0.0 now, because freshly created data frames created through `vec_init()` will have columns that all have a refcount of 1. The increment of the refcount happens when the columns are assigned into the data frame with `SET_VECTOR_ELT()`.

Practically this makes a difference in `vec_rbind()` where we assign into a fresh data frame in a loop, repeatedly calling `df_assign()`. Since each column is already _referenced_, the old scheme with `r_maybe_duplicate()` would duplicate the column at every iteration of the rbind loop. The new scheme allows us to take advantage of `owned = true` here because each column is referenced _but not shared_.

The usage of `owned = true` should be done incredibly sparingly. It should only be done when you recursively own the data structure you just created, and are filling it in some way.