@DavisVaughan 2020-04-29
I do believe we will have to expose the `owned = true` case in the public API in some way. In slider, I use `vec_proxy_assign()` in a loop to fill an `out` object that I internally create with `vec_init()`. This is exactly the same kind of case as `vec_rbind()`, where performance can be crippled if `out` is a data frame.

For an example, when `slide_vec(.ptype = <data.frame>)` is used, performance suffers a lot from this on R 4.0.0.

```r
library(slider)
library(purrr)

df <- data.frame(x = 1:5000, y = 1:5000)
```

```r
# Master / R 3.6
bench::mark(slide_vec(df, identity, .ptype = df))
#> # A tibble: 1 x 6
#>   expression                              min median `itr/sec` mem_alloc
#>   <bch:expr>                           <bch:> <bch:>     <dbl> <bch:byt>
#> 1 slide_vec(df, identity, .ptype = df) 14.3ms 17.4ms      58.4     291KB
#> # … with 1 more variable: `gc/sec` <dbl>

# This PR + R 4.0.0
bench::mark(slide_vec(df, identity, .ptype = df))
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 slide_vec(df, identity, .ptype = df)  79ms   84ms      11.2     191MB     44.8
```

Performance doesn't suffer with non-data frame objects because their reference count is `0` when they are created with `vec_init()`. Only the columns of a freshly created data frame have the issue.

```r
# Master / R 3.6
bench::mark(slide_int(df$x, identity), map_int(df$x, identity), iterations = 200)
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 slide_int(df$x, identity)   2.21ms    2.8ms      353.    29.5KB     28.6
#> 2 map_int(df$x, identity)     1.97ms   2.22ms      445.    33.7KB     30.9

# This PR + R 4.0.0
bench::mark(slide_int(df$x, identity), map_int(df$x, identity), iterations = 200)
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 slide_int(df$x, identity)   2.57ms   3.08ms      331.    29.5KB     24.9
#> 2 map_int(df$x, identity)     2.16ms   2.43ms      409.    33.6KB     33.2
```