@DavisVaughan 2020-04-29
I should also add that the original reason we went away from the `clone` parameter was outlined in this PR https://github.com/r-lib/vctrs/pull/842

It had to do with ALTREP, where `vec_init()` would construct a data frame with ALTREP compact-rep columns. We previously had `x = clone ? r_maybe_duplicate(x) : x;` which was a problem with this approach because in `vec_rbind()` we would set `clone = false`. This means that we would be incorrectly modifying ALTREP columns directly. They are marked as mutable so they should have been duplicated. This approach avoids that problem because we use `r_maybe_duplicate_shared(x)` even when we own `x`. ALTREP columns _are_ considered shared because by marking them as mutable we set their refcount to the max value, so they are correctly duplicated here.