@DavisVaughan 2020-04-30
An additional note that I think continuing to use `MAYBE_REFERENCED()` (i.e. refcount > 0) to determine if we need to duplicate is the correct thing to do in 99% of the cases. It matches up with this note from Luke in https://developer.r-project.org/Refcnt.html

"This means that, except in very special and well understood circumstances, an argument passed down to C code should not be modified if it has a positive reference count, even if that count is equal to one."

So I think https://github.com/r-lib/vctrs/issues/838 could be closed because of this